Added the In Memory SQL DAO and altered some of the models
Getting used to using Github. Added in the In Memory SQL DAO and altered some of the models to fit the implementation. Added sample data to tags, nodes, node tags, and user ratings.

Altered the Create Table sql file to have accounts have primary key of username and authorization level. Removed rating as a rating is something done by a separate user.